fix(mcp): GLM MCP 服务兼容性修复
- 添加 Accept 头避免 406 错误
- 支持 SSE 格式响应解析
- 新增 Mcp-Session-Id 管理和传递
- 发送 notifications/initialized 通知
- 所有请求添加 JSON-RPC ID 避免 nil pointer
- SyncTools 自动重连检查
- sync 端点从 DB 加载缺失的 client
- 前端 IconSync → IconRefresh 修复
- MCPToolStats.AvgDuration 类型改为 float64

+++ one-api/model/mcp_log.go
@@ -59,17 +59,17 @@ type MCPProviderStats struct {
 	TotalCalls     int64  `json:"total_calls"`
 	SuccessCalls   int64  `json:"success_calls"`
 	FailedCalls    int64  `json:"failed_calls"`
-	AvgDuration    int64  `json:"avg_duration"`    // ms
-	LastCalledAt   int64  `json:"last_called_at"`
+	AvgDuration    float64 `json:"avg_duration"`    // ms
+	LastCalledAt   int64   `json:"last_called_at"`
 }
 
 // MCPToolStats holds aggregated stats for a tool.
 type MCPToolStats struct {
-	ToolName     string `json:"tool_name"`
-	TotalCalls   int64  `json:"total_calls"`
-	SuccessCalls int64  `json:"success_calls"`
-	FailedCalls  int64  `json:"failed_calls"`
-	AvgDuration  int64  `json:"avg_duration"`
+	ToolName     string  `json:"tool_name"`
+	TotalCalls   int64   `json:"total_calls"`
+	SuccessCalls int64   `json:"success_calls"`
+	FailedCalls  int64   `json:"failed_calls"`
+	AvgDuration  float64 `json:"avg_duration"`
 }
 
 // GetMCPProviderStats returns aggregated stats grouped by provider.
